Branch for those who want Ian Schofield's incandescent lamp simulator code backed out.  It is not simply a reversion to [3a85f08639], it still contains all of the improvements we've made on the trunk.
This branch will continue to be periodically refreshed from the trunk
until we manage to extract the front panel code from the simulator, at
which point we'll need to take a second stab at making this code
optional. At that time, it will be possible to disable it with a
configure script option rather than make the user switch branches in the
Fossil repo.

FossilOrigin-Name: cb64c4a4b26ae7648a459cc23db6697e0e4cd919