All of the boot scripts except for those generated from assembly code now have a SIMH throttle value set based on either detected CPU cores at configuration time or the value of the new --throttle configure script value.  Added README-throttle.md to describe the values allowed for this option, as they're pretty extensive.
On a multi-core machine, the default is to have the prior behavior,
which is no throttling: run flat-out, taking up a whole CPU core for
pidp8i-sim's main thread, leaving the rest for the GPIO thread and
whatever else the Pi does.

On a single-core machine, the default is to set a value that should give
something like 50% CPU usage to pidp8i-sim on the slowest Raspberry Pis
we support, allowing enough CPU power left over for the incandescent
lamp simulator thread.  Consequently, removed README-single-core.md and
the references to it.

Closes [dfd864da90].

FossilOrigin-Name: 48b84d4bd947234851648732eb76b4c137fd184f