First attempt at taking the simple bitfields used for communicating register values between the PDP-8 CPU instruction execution loop and the PiDP-8/I front panel GPIO thread and switching them to a 2D array of LED "on" counts, allowing us to accuractely calculate the brightness of the LED at that time.
This code builds on macOS, but it has yet to be tested on the Pi, which
is part of why this is being checked in on a branch.  The other reason
is that we're pretty sure it's going to take at least one more checkin
to turn this into something we can ship.

FossilOrigin-Name: 66a2818a47b1d8c64de3d51413e573e5550b4d0b