Changed tools/version to use 10 hex digits for the Fossil checkin ID part to match the Fossil timeline default length, to avoid confusion when telling people to match the value given here to the value on the timeline page.  Not everyone will understand that 8 hex digits is more than enough to make a unique match in our small-ish repository.
Also made the build depend on tools/version such that when that script
file changes, it does a rebuild in case some future change like this
occurs.  It wasn't necessary this time because we also touched
Makefile.in, but we might not always do so in the future.

FossilOrigin-Name: b68d4a68fc84bee25722748fa7308c6e5e7233c2