Rather than duplicate all of the option definitions and help texts between auto.def and lib/mkos8/argparser.py, we now generate lib/mkos8/opts.py from auto.def based on the options() values at the top of auto.def which mkos8's argument parser reads and uses to define most of its command line options, making a near-parallel of most of the configure script's --*-os8-*.
This is on a branch because there are quite a few subtleties in all of
this which means this feature needs more testing to be sure it is
complete and doesn't break anything.

So far, this change is a delta loss of 132 lines, a good sign that it's
worth the complexity of doing things this way.

Provisionally closes [f5d5cf50c4] and [68dae7cc88].

FossilOrigin-Name: 9dcaed3c6447fa2e7d94c7bf313bdff68501fe9b