Added support for Pi 4 GPIO, based on code the Raspberry Pi Foundation published: https://github.com/RPi-Distro/raspi-gpio/commit/80fa7d04eafb3ea34fc6f2d32de5f1873b5fb369
This is a branch because it conflicts with the current tip-of-trunk and
because it's not yet tested on a Pi 4.

It is unkonwn whether this branch will end up overriding the method
we're using on trunk for Pi 0-3 or if we will instead wait for the
Foundation to publish a new libbcm_host that obviates some of what this
check-in does.

FossilOrigin-Name: bd10d4dde2f485a510a7eddd44f9a99ce09997eaa2aa7edc52183d207943849e